doc-qa-interface/requirements.txt
fastapi
uvicorn
langchain
boto3
faiss-cpu
numpy
anthropic
